+++ classes/rental.rb
@@ -1,5 +1,5 @@
-require './book'
-require './person'
+require_relative 'book'
+require_relative 'person'
 
 class Rental
   attr_accessor :date, :book, :person
