+++ decorators/capitalize.rb
@@ -1,4 +1,4 @@
-require './decorator'
+require_relative 'decorator'
 
 class CapitalizeDecorator < Decorator
   def correct_name
