+++ decorators/decorator.rb
@@ -1,4 +1,4 @@
-require './nameable'
+require_relative 'classes/nameable'
 
 class Decorator < Nameable
   attr_accessor :nameable
