+++ main.rb
@@ -1,4 +1,4 @@
-require './app'
+require_relative 'app'
 
 def main
   app = App.new
